Assignment: Recognise unquoted rvalue variable
As already mentioned in commit "Assignment: Fixed and simplified",
in the assignment 'foo=$1; shift', the variable is not tagged correctly,
even though :rvalue includes :variable. The same holds for just 'foo=$1'.

The reason for this is a little bit tricky:

Each variable pattern ends with a positive lookahead for a non-word
character, wich is absolutely sensible since it ensures that variables
are tagged correctly in an expression like "$basePath/$targetDir".

Now let's take a look at 'foo=$1' (followed by a newline): Our rvalue
pattern captures everything after the '=', but not the newline.
Thus, '$1' is matched against the variable patterns, and since there
is nothing after the variable, the assertion for a trailing non-word
character fails.

Regarding 'foo=$1; shift', you might think that there IS a trailing
non-word character, but here, our rvalue pattern (of course) doesn't
capture the ';', thus also in this case, '$1' is matched against
the variable patterns.

So to take into account that a variable name may be followed by nothing,
we first look ahead for /\W/ and if that fails, we look ahead for /$/.