textmate_tools/regexp.rb: "Optimise" code, but I think this introduced a bug.
ATM I don't have the time to investigate which changes are the bad ones,
so I just commit them and then revert the commit.

EDIT: This commit was indeed the root of evil.
After reverting it, my splitted and revised
TextMate tools produce the same output as
Jeff's original single-file version.
With the changes introduced by this commit,
there were no non-capture groups to wrap alternations.
This way, of course, nothing works.